# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
  ActiveBooking:
    properties:
      floor:
        type: integer
      room_number:
        type: integer
    type: object
...
  Assistance:
    properties:
      accessibility:
        items:
          type: string
        type: array
      dietary:
        items:
          type: string
        type: array
      medical:
        items:
          type: string
        type: array
    type: object
  CreateGuest:
    properties:
      first_name:
        example: Jane
        type: string
      last_name:
        example: Doe
        type: string
      profile_picture:
        example: https://example.com/john.jpg
        type: string
      timezone:
        example: America/New_York
        type: string
    type: object
...
  FilterRoomsRequest:
    properties:
      advanced:
        description: arrivals-today | departures-today
        items:
          type: string
        type: array
      attributes:
        description: standard | deluxe | suite | accessible
        items:
          type: string
        type: array
      cursor:
        type: string
      floors:
        items:
          type: integer
        type: array
      limit:
        minimum: 0
        type: integer
      sort:
        $ref: '#/definitions/github_com_generate_selfserve_internal_models.RoomSortOption'
      status:
        description: occupied | vacant | open-tasks
        items:
          type: string
        type: array
    type: object
...
  Guest:
    properties:
      created_at:
        example: "2024-01-02T00:00:00Z"
        type: string
      first_name:
        example: Jane
        type: string
      id:
        example: 530e8400-e458-41d4-a716-446655440000
        type: string
      last_name:
        example: Doe
        type: string
      notes:
        example: VIP guest
        type: string
      profile_picture:
        example: https://example.com/john.jpg
        type: string
      timezone:
        example: America/New_York
        type: string
      updated_at:
        example: "2024-01-02T00:00:00Z"
        type: string
    type: object
  GuestFilters:
    properties:
      assistance:
        items:
          $ref: '#/definitions/github_com_generate_selfserve_internal_models.AssistanceFilter'
        type: array
      cursor:
        type: string
      floor_sort:
        allOf:
        - $ref: '#/definitions/github_com_generate_selfserve_internal_models.FloorSortOrder'
        enum:
        - ascending
        - descending
      floors:
        items:
          type: integer
        type: array
      group_size:
        items:
          type: integer
        type: array
      limit:
        maximum: 100
        minimum: 1
        type: integer
      request_sort:
        allOf:
        - $ref: '#/definitions/github_com_generate_selfserve_internal_models.RequestSortOrder'
        enum:
        - high_to_low
        - low_to_high
        - urgent
      search:
        type: string
      status:
        items:
          $ref: '#/definitions/github_com_generate_selfserve_internal_models.BookingStatus'
        type: array
    type: object
  GuestPage:
    properties:
      data:
        items:
          $ref: '#/definitions/GuestWithBooking'
        type: array
      next_cursor:
        type: string
    type: object
...
  GuestWithBooking:
    properties:
      active_bookings:
        items:
          $ref: '#/definitions/ActiveBooking'
        type: array
      assistance:
        $ref: '#/definitions/Assistance'
      first_name:
        type: string
      has_urgent:
        type: boolean
      id:
        type: string
      last_name:
        type: string
      preferred_name:
        type: string
      request_count:
        type: integer
    type: object
  GuestWithStays:
    properties:
      assistance:
        $ref: '#/definitions/Assistance'
      current_stays:
        items:
          $ref: '#/definitions/Stay'
        type: array
      do_not_disturb_end:
        example: "07:00:00"
        type: string
      do_not_disturb_start:
        example: "17:00:00"
        type: string
      email:
        example: jane.doe@example.com
        type: string
      first_name:
        example: Jane
        type: string
      housekeeping_cadence:
        example: daily
        type: string
      id:
        example: 530e8400-e458-41d4-a716-446655440000
        type: string
      last_name:
        example: Doe
        type: string
      notes:
        example: VIP
        type: string
      past_stays:
        items:
          $ref: '#/definitions/Stay'
        type: array
      phone:
        example: +1 (617) 012-3456
        type: string
      preferences:
        example: extra pillows
        type: string
      pronouns:
        example: she/her
        type: string
    required:
    - current_stays
    - first_name
    - id
    - last_name
    - past_stays
    type: object
...
  RoomWithOptionalGuestBooking:
    properties:
      booking_status:
        $ref: '#/definitions/github_com_generate_selfserve_internal_models.BookingStatus'
      floor:
        type: integer
      guests:
        items:
          $ref: '#/definitions/Guest'
        type: array
      has_unassigned_tasks:
        type: boolean
      id:
        type: string
      is_accessible:
        type: boolean
      priority:
        $ref: '#/definitions/github_com_generate_selfserve_internal_models.RequestPriority'
      room_number:
        type: integer
      room_status:
        type: string
      suite_type:
        type: string
    type: object
  Stay:
    properties:
      arrival_date:
        example: "2024-01-02"
        type: string
      departure_date:
        example: "2024-01-05"
        type: string
      group_size:
        type: integer
      room_number:
        example: 101
        type: integer
      status:
        $ref: '#/definitions/github_com_generate_selfserve_internal_models.BookingStatus'
    required:
    - arrival_date
    - departure_date
    - room_number
    - status
    type: object
  UpdateGuest:
    properties:
      first_name:
        example: Jane
        type: string
      last_name:
        example: Doe
        type: string
      notes:
        example: VIP guest
        maxLength: 1000
        type: string
      profile_picture:
        example: https://example.com/john.jpg
        type: string
      timezone:
        example: America/New_York
        type: string
    type: object
...
  github_com_generate_selfserve_internal_errs.HTTPError:
    properties:
      code:
        type: integer
      message: {}
    type: object
  github_com_generate_selfserve_internal_models.AssistanceFilter:
    enum:
    - accessibility
    - dietary
    - medical
    type: string
    x-enum-varnames:
    - AssistanceAccessibility
    - AssistanceDietary
    - AssistanceMedical
  github_com_generate_selfserve_internal_models.BookingStatus:
    enum:
    - active
    - inactive
    type: string
    x-enum-varnames:
    - BookingStatusActive
    - BookingStatusInactive
...
  github_com_generate_selfserve_internal_models.FloorSortOrder:
    enum:
    - ascending
    - descending
    type: string
    x-enum-varnames:
    - FloorSortAscending
    - FloorSortDescending
...
  github_com_generate_selfserve_internal_models.RequestPriority:
    enum:
    - low
    - medium
    - high
    type: string
    x-enum-varnames:
    - PriorityLow
    - PriorityMedium
    - PriorityHigh
  github_com_generate_selfserve_internal_models.RequestSortOrder:
    enum:
    - high_to_low
    - low_to_high
    - urgent
    type: string
    x-enum-varnames:
    - RequestSortHighToLow
    - RequestSortLowToHigh
    - RequestSortUrgent
  github_com_generate_selfserve_internal_models.RoomSortOption:
    enum:
    - ascending
    - descending
    - urgency
    type: string
    x-enum-varnames:
    - RoomSortAscending
    - RoomSortDescending
    - RoomSortUrgency
...
  github_com_generate_selfserve_internal_utils.CursorPage-RoomWithOptionalGuestBooking:
    properties:
      has_more:
        type: boolean
      items:
        items:
          $ref: '#/definitions/RoomWithOptionalGuestBooking'
        type: array
      next_cursor:
        description: nil when no more pages
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@selfserve.com
    name: API Support
  description: SelfServe backend API documentation
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: SelfServe API
  version: "1.0"
paths:
...
  /guests:
    post:
      consumes:
      - application/json
      description: Creates a guest with the given data
      parameters:
      - description: Guest data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/CreateGuest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Guest'
        "400":
          description: Invalid guest body format
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Creates a guest
      tags:
      - guests
  /guests/{id}:
    get:
      consumes:
      - application/json
      description: Retrieves a single guest given an id
      parameters:
      - description: Guest ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Guest'
        "400":
          description: Invalid guest ID format
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Guest not found
          schema:
            $ref: '#/definitions/github_com_generate_selfserve_internal_errs.HTTPError'
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Gets a guest
      tags:
      - guests
    put:
      consumes:
      - application/json
      description: Updates fields on a guest
      parameters:
      - description: Guest ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: Guest update data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/UpdateGuest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Guest'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Updates a guest
      tags:
      - guests
  /guests/search:
    post:
      consumes:
      - application/json
      description: Retrieves guests optionally filtered by floor
      parameters:
      - description: Hotel ID (UUID)
        in: header
        name: X-Hotel-ID
        required: true
        type: string
      - description: Guest filters
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/GuestFilters'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/GuestPage'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get Guests
      tags:
      - guests
  /guests/stays/{id}:
    get:
      consumes:
      - application/json
      description: Retrieves a single guest with previous stays given an id
      operationId: getGuestsStaysId
      parameters:
      - description: Guest ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/GuestWithStays'
        "400":
          description: Invalid guest ID format
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Guest not found
          schema:
            $ref: '#/definitions/github_com_generate_selfserve_internal_errs.HTTPError'
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Gets a guest with previous stays
      tags:
      - guests
...
  /rooms:
    post:
      consumes:
      - application/json
      description: Retrieves rooms with optional filters (status, attributes, advanced,
        sort) and cursor pagination. Occupancy is derived from active guest_bookings;
        open tasks from the latest request version per room.
      parameters:
      - description: Hotel ID
        in: header
        name: X-Hotel-ID
        required: true
        type: string
      - description: Filters and pagination
        in: body
        name: body
        schema:
          $ref: '#/definitions/FilterRoomsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_generate_selfserve_internal_utils.CursorPage-RoomWithOptionalGuestBooking'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List rooms with filters
      tags:
      - rooms
  /rooms/{id}:
    get:
      description: Retrieves a single room by its UUID
      parameters:
      - description: Hotel ID (UUID)
        in: header
        name: X-Hotel-ID
        required: true
        type: string
      - description: Room ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/RoomWithOptionalGuestBooking'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_generate_selfserve_internal_errs.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_generate_selfserve_internal_errs.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_generate_selfserve_internal_errs.HTTPError'
      security:
      - BearerAuth: []
      summary: Get room by ID
      tags:
      - rooms
  /rooms/floors:
    get:
      description: Retrieves all distinct floor numbers
      parameters:
      - description: Hotel ID (UUID)
        in: header
        name: X-Hotel-ID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              type: integer
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get Floors
      tags:
      - rooms
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Your Clerk JWT token (prefixed with "Bearer ")
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"